Fix taglist to work properly when clicking on an autocomplete option
The taglist was setting a tag as soon as an option was clicked but
before the model was updated with the full value of the autocompleted
tag. So, we would end up with the partial text as the tag and the
autocompleted text in the input box. The taglist directive was updated
with an on-blur timeout for this very reason, the fix was to use it when
doing autocomplete.