Show an icon for rerun transactions on list and details pages
Prior to this change, there was no way to show the user that a
transaction is the result of a rerun transaction request.

This change will show an icon (glyphicon-retweet) for transactions
that are reruns in the 'transactions log' page. It will also show
an icon for these transactions on the 'transaction details' page.
Where the transaction is a reun and the parent of a rerun, both
indicators will be displayed.

OHM-662