Fix role deselection on edit client modal
Prior to this change, initially clicking a role checkbox on the
'Edit Client' modal would uncheck the box, but not remove the role
from the client. Subsequent clicks on the same role checkbox, would
correctly tick/untick the checkbox and update the client role
appropriately i.e. the initial click on the checkbox was not
triggering the ng-change event

The ng-click always fires regardless of the checkbox state,
allowing the app to update the roles every time the checkbox
state changes.

OHM-747
https://stackoverflow.com/questions/39905845/ng-change-is-not-working-after-one-selection