wrap the body range properties in an object that will be passed to controllers
The range prperties will be passed on to controllers that display the body. This is so the controller for those components knows how much of the body has been retrieved and whether to display components for retrieving the full body. The Api is now also being passedd into the TransactionBodyModalCtrl

OHM-1039