modify the transactionBodyCtrl
The body range properties are now being passed in and the Api service is not being passed in as an argument. The function for retriving the body in now only called on request when more of the body has to be retrieved. The body range properties being passed in will be used to determine whether components for displaying more of the body are rendered.

OHM-1039