add the body range properties to the TransctionAddReqResModalCtrl
The body range properties are used to detemine whether the whole body has been passed in, and whether to render components for retrieving the body (if partial). The controller has also been refactored. The call to retrieve the bosy will only be done whe the reqeust body has not been passed

OHM-1039